# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda API - Documentação",
        "version": "0.1"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login de usuário",
                "operationId": "99d66635c4992aeaa6aa44ff653d0563",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "example@teste.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário autenticado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status da solicitação",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "description": "Token de acesso",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status da solicitação",
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrar novo usuário",
                "operationId": "08136088c4862c313b35d6518bab8d3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status da solicitação",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obter informações do usuário autenticado",
                "operationId": "32d03117b496ae02c2543b61438d24ad",
                "responses": {
                    "200": {
                        "description": "Informações do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status da solicitação",
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/calendar-types": {
            "get": {
                "tags": [
                    "Calendar Types"
                ],
                "summary": "Listar tipos de calendário",
                "operationId": "52044a95d8d924d5b3a883c551932b26",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de calendário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CalendarType"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Calendar Types"
                ],
                "summary": "Criar tipo de calendário",
                "operationId": "aa6ef8cf12272a06b2093de56409dfb5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CalendarTypePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de calendário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalendarType"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/calendar-types/{id}": {
            "get": {
                "tags": [
                    "Calendar Types"
                ],
                "summary": "Obter tipo de calendário",
                "operationId": "a8ec95b0dd63d619a6454cc2848de7a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de calendário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de calendário obtido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalendarType"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Calendar Types"
                ],
                "summary": "Deletar tipo de calendário",
                "operationId": "deaab3c19575d64784f4a57a0bd480cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de calendário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de calendário deletado com sucesso"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Calendar Types"
                ],
                "summary": "Atualizar tipo de calendário",
                "operationId": "d50d7df0aaaa4442ab0a8c43e5e5bdec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do tipo de calendário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CalendarTypePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de calendário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CalendarType"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CalendarTypePostRequest": {
                "title": "Calendar Type Post Request",
                "description": "Schema de criação de calendar type.",
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "description": "Nome do Calendar Type",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do Calendar Type",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserPostRequest": {
                "title": "User Post Request",
                "description": "Schema de criação de usuário.",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "E-mail do usuário",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "Senha do usuário (mínimo de 8 caracteres)",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
...
            "CalendarType": {
                "title": "CalendarType",
                "description": "Modelo de tipo de calendário",
                "properties": {
                    "id": {
                        "description": "ID do tipo de calendário",
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "description": "Nome do tipo de calendário",
                        "type": "string"
                    },
                    "description": {
                        "description": "Descrição do tipo de calendário",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID do usuário proprietário do tipo de calendário",
                        "type": "integer",
                        "format": "int64"
                    },
                    "created_at": {
                        "description": "Data de criação do tipo de calendário",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização do tipo de calendário",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Modelo de usuário",
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string"
                    },
                    "email": {
                        "description": "Endereço de email do usuário",
                        "type": "string",
                        "format": "email"
                    },
                    "created_at": {
                        "description": "Data de criação do usuário",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização do usuário",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}